fix: CustomTypeMapping の分割
