+++ core/src/test/scala/com/iheart/playSwagger/DefinitionGeneratorSpec.scala
@@ -1,6 +1,7 @@
 package com.iheart.playSwagger
 
 import com.iheart.playSwagger.Domain._
+import com.iheart.playSwagger.domain.CustomTypeMapping
 import org.specs2.mutable.Specification
 import play.api.libs.json.Json
